requirements.txt
absl-py==2.2.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.47.2
anyascii==0.3.2
anyio==4.8.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==25.1.0
audioread==3.0.1
av==10.0.0
babel==2.17.0
backoff==2.2.1
bangla==0.0.2
bcrypt==4.3.0
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.37.25
botocore==1.37.25
build==1.2.2.post1
cached_path==1.7.1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cloudpathlib==0.21.0
cn2an==0.5.22
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
ctranslate2==3.24.0
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dateparser==1.1.8
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.18
deptry==0.23.0
Distance==0.1.3
distro==1.9.0
docker==7.1.0
docker-pycreds==0.4.0
docopt==0.6.2
dotenv==0.9.9
dtw-python==1.4.4
durationpy==0.9
einops==0.8.1
encodec==0.1.1
eng_to_ipa==0.0.2
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.12
faster-whisper==0.9.0
ffmpy==0.5.0
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
fugashi==1.3.0
funcy==2.0
g2p-en==2.1.0
g2pkk==0.1.2
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gradio==3.48.0
gradio_client==0.6.1
greenlet==3.1.1
groq==0.18.0
grpcio==1.67.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.14.0
hangul-romanize==0.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflect==7.0.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.19
langchain-anthropic==0.3.8
langchain-core==0.3.40
langchain-groq==0.2.4
langchain-huggingface==0.1.2
langchain-milvus==0.1.8
langchain-ollama==0.2.3
langchain-openai==0.3.7
langchain-postgres==0.0.13
langchain-text-splitters==0.3.6
langcodes==3.5.0
langgraph==0.3.2
langgraph-checkpoint==2.0.16
langgraph-prebuilt==0.1.1
langgraph-sdk==0.1.53
langid==1.1.6
langsmith==0.3.11
language_data==1.3.0
lazy_loader==0.4
librosa==0.9.1
llvmlite==0.43.0
loguru==0.7.2
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mecab-python3==1.0.9
melotts @ git+https://github.com/myshell-ai/MeloTTS.git@209145371cff8fc3bd60d7be902ea69cbdb7965a
milvus-lite==2.4.12
mmh3==5.1.0
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
narwhals==1.29.0
nest-asyncio==1.6.0
networkx==2.8.8
nltk==3.9.1
num2words==0.5.12
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.7
onnxruntime==1.21.0
openai==1.64.0
openai-whisper==20240930
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.15
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.1
pandas==1.5.3
parso==0.8.4
pexpect==4.9.0
pgvector==0.3.6
pillow==10.4.0
plac==1.4.3
platformdirs==4.3.2
pooch==1.8.2
posthog==3.23.0
preshed==3.0.9
proces==0.1.7
prompt_toolkit==3.0.47
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==6.0.0
psycopg==3.2.5
psycopg-binary==3.2.5
psycopg-pool==3.2.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pykakasi==2.2.1
pymilvus==2.5.4
pynndescent==0.5.13
pyparsing==3.2.3
PyPika==0.48.9
pypinyin==0.50.0
pyproject_hooks==1.2.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requirements-parser==0.11.0
resampy==0.4.3
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.11.4
semantic-version==2.10.0
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.1
streamlit==1.42.2
SudachiDict-core==20250129
SudachiPy==0.6.10
suno-bark @ git+https://github.com/suno-ai/bark.git@f4f32d4cd480dfec1c245d258174bc9bde3c2148
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
tornado==6.4.1
tqdm==4.67.1
trainer==0.0.36
traitlets==5.14.3
transformers==4.50.3
TTS==0.22.0
txtsplit==1.0.0
typeguard==4.4.2
typer==0.15.2
types-setuptools==78.1.0.20250329
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
ujson==5.10.0
umap-learn==0.5.7
Unidecode==1.3.7
unidic==1.1.0
unidic-lite==1.0.8
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
wasabi==0.10.1
watchdog==6.0.0
watchfiles==1.0.4
wavmark==0.0.3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
whisper-timestamped==1.14.2
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
